ref: `profile` -> `hostAttrs`
Rationale:

For internal coding purposes only the following attribute set does
not require an abstract name (such as `profile`). Moreover so, as
"profile" has a distinct legacy meaning in the nix ecosystem.
 {
    channelName = "...";
    system = "...";
    extraArgs = {};
    modules = [];
 };

Instead, this PR renames it to a banal and decribtive `hostAttrs`. This
naming does also better aling with the top level `defaultHostAttrs`.