add: lib.builders.packagesFromOverlaysBuiderConstructer
Constructor, that with the help of channels dependency constructs a
packagesFromOverlaysBuilder that outputs all packages defined
in overlays that do not break flake checks.

Good for sharing binary cached packages. Any package that would
break nix flake check (e.g. because it is marked as broken) cannot
be shared via this mechanism, and the alternative via overlays has to be
used.