Revert patching functionality and fix config handling
The patching functionality is fundamentally ill designed
and makes the code complex to the point that it is increasingly
hard to audit it for correctness.

The root of evil and ill-design is that it is a work-around for
what the Unoficial Flake Roadmap describes as follows:
- flake inputs are flake inputs
- patched flake inputs are flake inputs

Furthermore this commit fixes a naive implementation of nixpkgs.config
handling that causes nasty regressions on edge cases. Namely, not only
can a nixpkgs config be a function, but also the merging semantics
as defined by the module system requires detailed knowledge of the
potential shape of the cofig attrs / function.

For those users that have been accustomed to the patching facility
and do not suffer from the regression-causing old implementation
of nixpkgs config handling, I set up
https://github.com/gytis-ivaskevicius/flake-utils-plus/tree/staging-with-nixpkgs-patching
which reverts this commit and can be continously rebased until
it will be no longer necesary as upstream inputs patching will
be implemented.